[*] move functions of console in cpp, made them static
