--- lib/class/Console.cpp
@@ -1,1 +1,0 @@
-#include "../../includes/class/Console.hpp"
